- splits typescript writer in visitor pattern
