+++ src/Kiota.Builder/Writers/TypeScript/CodeClassEndWriter.cs
@@ -1,0 +1,10 @@
+namespace  Kiota.Builder.Writers.TypeScript {
+    public class CodeClassEndWriter : ICodeElementWriter<CodeClass.End>
+    {
+        public void WriteCodeElement(CodeClass.End codeElement, LanguageWriter writer)
+        {
+            writer.DecreaseIndent();
+            writer.WriteLine("}");
+        }
+    }
+}
